Added run, iteration, and trial functions.
Run: performs a single run of a given commander solution and returns the final score.
Iteration: Calls for n runs of a given commander solution and returns the mean final score.
Trial: Calls for n runs of unique commander solutions, retaining the best performing and returns the best solution and its mean score.